# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mama Health API
  version: 1.0.0
  description: REST API for the Mama Health pregnancy care platform. Serves the Patient,
    Doctor, and Admin (Flutter Web) clients.
paths:
...
  /api/v1/appointments/:
    get:
      operationId: appointments_list
      description: 'Role-scoped: patient sees only their own; doctor sees only their
        own (not other doctors'' appointments with a shared patient) — `?patient_id=`/`?doctor_id=`
        have no effect for a doctor caller. Admin sees all, optionally narrowed with
        `?patient_id=` and/or `?doctor_id=`. Paginated.'
      summary: List appointments
      parameters:
      - in: query
        name: doctor_id
        schema:
          type: integer
        description: 'Admin only: narrow to one doctor''s appointments.'
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: patient_id
        schema:
          type: integer
        description: 'Admin only: narrow to one patient''s appointment history.'
      tags:
      - Appointments
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAppointmentList'
              examples:
                200OK:
                  value:
                    count: 123
                    next: http://api.example.org/accounts/?page=4
                    previous: http://api.example.org/accounts/?page=2
                    results:
                    - count: 1
                      next: null
                      previous: null
                      results:
                      - id: 15
                        patient:
                          id: 2
                          email: sara.ahmed@example.com
                          first_name: Sara
                          last_name: Ahmed
                        doctor:
                          id: 4
                          email: dr.ayesha@example.com
                          first_name: Ayesha
                          last_name: Malik
                        appointment_type: in_person
                        scheduled_at: '2026-08-05T10:30:00Z'
                        duration_minutes: 30
                        status: pending
                        meeting_link: ''
                        reason: Routine 20-week checkup
                        doctor_notes: ''
                        cancellation_reason: ''
                        created_at: '2026-07-29T14:00:00Z'
                        updated_at: '2026-07-29T14:00:00Z'
                  summary: 200 OK
          description: ''
    post:
      operationId: appointments_create
      description: Patients book for themselves — any `patient_id` in the payload
        is ignored/overridden with the requesting patient. A doctor/admin booking
        on behalf of a patient **must** supply `patient_id` (400 otherwise). `appointment_type=video_consultation`
        does **not** provision a real video session in v1 — `meeting_link` stays empty
        until manually filled in (no live video SDK integration). Booking auto-creates
        a `PatientDoctorAssignment` between this patient and doctor if one doesn't
        already exist, and notifies the doctor.
      summary: Book an appointment
      tags:
      - Appointments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentRequest'
            examples:
              Request—PatientBooking:
                value:
                  doctor_id: 4
                  appointment_type: in_person
                  scheduled_at: '2026-08-05T10:30:00Z'
                  duration_minutes: 30
                  reason: Routine 20-week checkup
                summary: Request — patient booking
              Request—Admin/doctorBookingOnBehalfOfAPatient:
                value:
                  patient_id: 2
                  doctor_id: 4
                  appointment_type: video_consultation
                  scheduled_at: '2026-08-05T10:30:00Z'
                summary: Request — admin/doctor booking on behalf of a patient
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AppointmentRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AppointmentRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
              examples:
                201Created:
                  value:
                    id: 15
                    patient:
                      id: 2
                      email: sara.ahmed@example.com
                      first_name: Sara
                      last_name: Ahmed
                    doctor:
                      id: 4
                      email: dr.ayesha@example.com
                      first_name: Ayesha
                      last_name: Malik
                    appointment_type: in_person
                    scheduled_at: '2026-08-05T10:30:00Z'
                    duration_minutes: 30
                    status: pending
                    meeting_link: ''
                    reason: Routine 20-week checkup
                    doctor_notes: ''
                    cancellation_reason: ''
                    created_at: '2026-07-29T14:00:00Z'
                    updated_at: '2026-07-29T14:00:00Z'
                  summary: 201 Created
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailResponse'
              examples:
                400MissingPatientId(non-patientActor):
                  value:
                    detail: Required when booking on behalf of a patient.
                    errors:
                      patient_id:
                      - Required when booking on behalf of a patient.
                  summary: 400 Missing patient_id (non-patient actor)
          description: ''
  /api/v1/appointments/{id}/:
    patch:
      operationId: appointments_partial_update
      description: |-
        Scoping is bespoke (not PatientScopedQuerysetMixin): an appointment has
        two distinct parties, so a doctor's view must be `doctor=request.user`,
        not "any patient this doctor is assigned to" (which would leak other
        doctors' appointments with a shared patient).
      summary: Update appointment fields (rarely used directly — prefer the /status/
        and /doctor-notes/ actions below)
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this appointment.
        required: true
      tags:
      - Appointments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAppointmentRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedAppointmentRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedAppointmentRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
          description: ''
    get:
      operationId: appointments_retrieve
      description: |-
        Scoping is bespoke (not PatientScopedQuerysetMixin): an appointment has
        two distinct parties, so a doctor's view must be `doctor=request.user`,
        not "any patient this doctor is assigned to" (which would leak other
        doctors' appointments with a shared patient).
      summary: Get a single appointment
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this appointment.
        required: true
      tags:
      - Appointments
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
              examples:
                200OK:
                  value:
                    id: 15
                    patient:
                      id: 2
                      email: sara.ahmed@example.com
                      first_name: Sara
                      last_name: Ahmed
                    doctor:
                      id: 4
                      email: dr.ayesha@example.com
                      first_name: Ayesha
                      last_name: Malik
                    appointment_type: in_person
                    scheduled_at: '2026-08-05T10:30:00Z'
                    duration_minutes: 30
                    status: pending
                    meeting_link: ''
                    reason: Routine 20-week checkup
                    doctor_notes: ''
                    cancellation_reason: ''
                    created_at: '2026-07-29T14:00:00Z'
                    updated_at: '2026-07-29T14:00:00Z'
                  summary: 200 OK
          description: ''
  /api/v1/appointments/{id}/doctor-notes/:
    patch:
      operationId: appointments_doctor_notes_partial_update
      description: Doctor/admin only (patients get 403). Free-text post-consultation
        notes, fully replaced (not appended) on each call.
      summary: Set doctor consultation notes
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this appointment.
        required: true
      tags:
      - Appointments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAppointmentDoctorNotesRequest'
            examples:
              Request:
                value:
                  doctor_notes: Blood pressure normal. Advised increased hydration.
                    Follow-up in 2 weeks.
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedAppointmentDoctorNotesRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedAppointmentDoctorNotesRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
              examples:
                200OK:
                  value:
                    id: 15
                    patient:
                      id: 2
                      email: sara.ahmed@example.com
                      first_name: Sara
                      last_name: Ahmed
                    doctor:
                      id: 4
                      email: dr.ayesha@example.com
                      first_name: Ayesha
                      last_name: Malik
                    appointment_type: in_person
                    scheduled_at: '2026-08-05T10:30:00Z'
                    duration_minutes: 30
                    status: pending
                    meeting_link: ''
                    reason: Routine 20-week checkup
                    doctor_notes: ''
                    cancellation_reason: ''
                    created_at: '2026-07-29T14:00:00Z'
                    updated_at: '2026-07-29T14:00:00Z'
                  summary: 200 OK
          description: ''
  /api/v1/appointments/{id}/reschedule/:
    patch:
      operationId: appointments_reschedule_partial_update
      description: 'Changes `scheduled_at` (and optionally `duration_minutes`) without
        going through the status state machine — status is untouched. Only allowed
        while `pending` or `confirmed` (400 for a terminal-status appointment: completed/cancelled/no_show).
        Notifies whichever party didn''t make the change (or both, if an admin made
        it), same as `/status/`.'
      summary: Reschedule an appointment
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this appointment.
        required: true
      tags:
      - Appointments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedAppointmentRescheduleRequest'
            examples:
              Request:
                value:
                  scheduled_at: '2026-08-10T14:00:00Z'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedAppointmentRescheduleRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedAppointmentRescheduleRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
              examples:
                200OK:
                  value:
                    id: 15
                    patient:
                      id: 2
                      email: sara.ahmed@example.com
                      first_name: Sara
                      last_name: Ahmed
                    doctor:
                      id: 4
                      email: dr.ayesha@example.com
                      first_name: Ayesha
                      last_name: Malik
                    appointment_type: in_person
                    scheduled_at: '2026-08-05T10:30:00Z'
                    duration_minutes: 30
                    status: pending
                    meeting_link: ''
                    reason: Routine 20-week checkup
                    doctor_notes: ''
                    cancellation_reason: ''
                    created_at: '2026-07-29T14:00:00Z'
                    updated_at: '2026-07-29T14:00:00Z'
                  summary: 200 OK
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailResponse'
              examples:
                400TerminalStatus:
                  value:
                    detail: Cannot reschedule a completed appointment.
                    errors: null
                  summary: 400 Terminal status
          description: ''
  /api/v1/appointments/{id}/status/:
    post:
      operationId: appointments_status_create
      description: 'State machine, enforced server-side — invalid transitions return
        400: `pending → confirmed | cancelled`, `confirmed → completed | cancelled
        | no_show`. `completed`/`cancelled`/`no_show` are terminal (no further transitions).
        `cancellation_reason` is only used/stored when `status=cancelled`. Notifies
        whichever party didn''t make the change (or both, if an admin made it).'
      summary: Change appointment status
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this appointment.
        required: true
      tags:
      - Appointments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentStatusUpdateRequest'
            examples:
              Request—Confirm:
                value:
                  status: confirmed
                summary: Request — confirm
              Request—Cancel:
                value:
                  status: cancelled
                  cancellation_reason: Patient has a scheduling conflict
                summary: Request — cancel
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AppointmentStatusUpdateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AppointmentStatusUpdateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
              examples:
                200OK:
                  value:
                    id: 15
                    patient:
                      id: 2
                      email: sara.ahmed@example.com
                      first_name: Sara
                      last_name: Ahmed
                    doctor:
                      id: 4
                      email: dr.ayesha@example.com
                      first_name: Ayesha
                      last_name: Malik
                    appointment_type: in_person
                    scheduled_at: '2026-08-05T10:30:00Z'
                    duration_minutes: 30
                    status: pending
                    meeting_link: ''
                    reason: Routine 20-week checkup
                    doctor_notes: ''
                    cancellation_reason: ''
                    created_at: '2026-07-29T14:00:00Z'
                    updated_at: '2026-07-29T14:00:00Z'
                  summary: 200 OK
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailResponse'
              examples:
                400InvalidTransition:
                  value:
                    detail: Cannot move an appointment from 'completed' to 'pending'.
                    errors: null
                  summary: 400 Invalid transition
          description: ''
...
components:
  schemas:
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        patient:
          allOf:
          - $ref: '#/components/schemas/BriefUser'
          readOnly: true
        doctor:
          allOf:
          - $ref: '#/components/schemas/BriefUser'
          readOnly: true
        appointment_type:
          $ref: '#/components/schemas/AppointmentTypeEnum'
        scheduled_at:
          type: string
          format: date-time
        duration_minutes:
          type: integer
          maximum: 2147483647
          minimum: 0
        status:
          allOf:
          - $ref: '#/components/schemas/AppointmentStatusEnum'
          readOnly: true
        meeting_link:
          type: string
          format: uri
          readOnly: true
        reason:
          type: string
        doctor_notes:
          type: string
        cancellation_reason:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - cancellation_reason
      - created_at
      - doctor
      - id
      - meeting_link
      - patient
      - scheduled_at
      - status
      - updated_at
    AppointmentRequest:
      type: object
      properties:
        doctor_id:
          type: integer
          writeOnly: true
        patient_id:
          type: integer
          writeOnly: true
        appointment_type:
          $ref: '#/components/schemas/AppointmentTypeEnum'
        scheduled_at:
          type: string
          format: date-time
        duration_minutes:
          type: integer
          maximum: 2147483647
          minimum: 0
        reason:
          type: string
        doctor_notes:
          type: string
      required:
      - doctor_id
      - scheduled_at
    AppointmentStatusEnum:
      enum:
      - pending
      - confirmed
      - completed
      - cancelled
      - no_show
      type: string
      description: |-
        * `pending` - Pending
        * `confirmed` - Confirmed
        * `completed` - Completed
        * `cancelled` - Cancelled
        * `no_show` - No show
    AppointmentStatusUpdateRequest:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/AppointmentStatusEnum'
        cancellation_reason:
          type: string
          default: ''
      required:
      - status
    AppointmentTypeEnum:
      enum:
      - in_person
      - video_consultation
      type: string
      description: |-
        * `in_person` - In person
        * `video_consultation` - Video consultation
...
    BriefUser:
      type: object
      description: |-
        Minimal user representation for nesting inside other apps' serializers
        (e.g. an Appointment's patient/doctor, a DietPlan's created_by).
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
      required:
      - email
      - id
...
    DetailResponse:
      type: object
      description: |-
        Generic {"detail": ..., "errors": ...} envelope — used purely to give
        non-default-status Swagger responses (201/400/403/404/503/etc.) a real
        schema entry so their OpenApiExample(status_codes=[...]) actually
        attaches instead of being silently dropped. Never used for real request
        validation, only as a `responses=` type hint for drf-spectacular.
      properties:
        detail:
          type: string
        errors:
          nullable: true
      required:
      - detail
...
    PaginatedAppointmentList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
...
    PatchedAppointmentDoctorNotesRequest:
      type: object
      properties:
        doctor_notes:
          type: string
    PatchedAppointmentRequest:
      type: object
      properties:
        doctor_id:
          type: integer
          writeOnly: true
        patient_id:
          type: integer
          writeOnly: true
        appointment_type:
          $ref: '#/components/schemas/AppointmentTypeEnum'
        scheduled_at:
          type: string
          format: date-time
        duration_minutes:
          type: integer
          maximum: 2147483647
          minimum: 0
        reason:
          type: string
        doctor_notes:
          type: string
    PatchedAppointmentRescheduleRequest:
      type: object
      properties:
        scheduled_at:
          type: string
          format: date-time
        duration_minutes:
          type: integer
          minimum: 1
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...